fix(scans): restore scan history listing + guard network-sweep tool choice
Scan Management showed "No scans found" (total>0, empty array) even though
scans ran and saved. Two independent bugs:

* list_scans SELECT had a duplicated "FROM scans FROM scans" in all four
  query variants (SQL syntax error).
* models::Scan.project_id was Option<i64> while the DB column is INTEGER
  (INT4). sqlx only type-checks non-NULL values, so rows with a project_id
  failed to decode; in fetch_all().unwrap_or_default() one bad row collapses
  the whole list to []. Matched the field to the column width (Option<i32>),
  which also fixes get_scan (SELECT *) silently 404ing on those rows.

Network sweep ran whatever tool was picked (e.g. gobuster) against every bare
IP, so web tools failed on every host and flooded history with useless
"failed" rows. Added a SWEEP_TOOLS allowlist (nmap/masscan/naabu/unicornscan/
fping/netdiscover), checked before the authorization gate with a clear message,
and the New Scan UI now only offers network scanners in Network Sweep mode.

+++ rust-backend/src/models/scan.rs
@@ -13,7 +13,11 @@ pub struct Scan {
     pub parameters: Option<JsonValue>,
     pub status: Option<String>,
     pub agent_id: Option<String>,
-    pub project_id: Option<i64>,
+    // DB column is INTEGER (INT4); sqlx decodes that to i32, not i64. Using i64
+    // here made every row with a non-NULL project_id fail to decode, which in a
+    // fetch_all() collapses the whole result to an empty Vec (list_scans showed
+    // total>0 but an empty scans array). Match the real column width.
+    pub project_id: Option<i32>,
     pub output: Option<String>,
     pub error_log: Option<String>,
     pub findings: Option<JsonValue>,
